+++ app/src/main/java/com/porfirio/orariprocida2011/threads/transports/DownloadTransportsHandler.java
@@ -1,4 +1,4 @@
-package com.porfirio.orariprocida2011.threads;
+package com.porfirio.orariprocida2011.threads.transports;
 
 import android.os.Bundle;
 import android.util.Log;
